refactor; no warnings on clang or gcc 5

+++ param.c
@@ -122,7 +122,7 @@ Param_Tweak(struct ocx *ocx, const char *arg)
 	}
 
 	assert (q >= arg);
-	l = (unsigned)(q - arg);
+	l = (size_t)(q - arg);
 
 	TAILQ_FOREACH(pt, &param_tbl, list) {
 		if (strlen(pt->name) != l)
